gesticulator/gesticulator/requirements.txt
absl-py==0.9.0
audioread==2.1.8
bvh==0.3
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
ConfigArgParse==1.2.3
cycler==0.10.0
decorator==4.4.2
docopt==0.6.2
fsspec==0.8.4
future==0.18.2
gluonnlp==0.6.0
google-auth==1.12.0
google-auth-oauthlib==0.4.1
graphviz==0.8.4
grpcio==1.27.2
h5py==2.10.0
idna==2.9
inflect==4.1.0
joblib==0.14.1
kiwisolver==1.1.0
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.2.1
matplotlib==3.2.1
nltk==3.5
numba==0.48.0
numpy==1.18.2
oauthlib==3.1.0
opensmile==1.0.1
pandas==1.0.3
praat-parselmouth==0.3.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.23.1
pyparsing==2.4.6
pyquaternion==0.9.5
pysptk==0.1.18
python-dateutil==2.8.1
python-speech-features==0.6
pytorch-crf==0.7.0
pytorch-lightning==1.1.0
pytorch-nlp==0.5.0
pytz==2019.3
PyYAML==5.3.1
regex==2020.10.28
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
SoundFile==0.10.3.post1
sox==1.4.1
stomp.py==6.1.0
transformers
torchtext==0.4.0
tqdm==4.44.1
transforms3d==0.3.1
typing==3.6.6
urllib3==1.25.8
Werkzeug==1.0.1
